Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
python-dotenv
alembic
websockets
pytest
httpx
